requirements.txt
Django==2.2.9
django-unicorn==0.22.0
Pillow
fpdf